完成所有ID格式的修改(Integer->Long), 完成回复功能的Controller层

+++ src/main/java/flower/community/mapper/UserMapper.java
@@ -17,7 +17,7 @@ public interface UserMapper {
     User findByToken(@Param("token") String token);
 
     @Select("SELECT * FROM USER WHERE ID = #{id}")
-    User findByID(@Param("id") Integer creator);
+    User findByID(@Param("id") Long creator);
 
     @Select("SELECT * FROM USER WHERE ACCOUNT_ID = #{accountId}")
     User findByAccountId(@Param("accountId") String accountId);
